Update pre-commit, linting and dev requirements
1. Disable autofix from pre-commit.ci

2. Add `nbstripout` in pre-commit which is a tool for stripping output from jupyter notebooks.
This is useful for keeping notebook size small and notebook diffs (semi-)sane.

3. Replace `prospector` with `ruff`. `ruff` is way faster and works great as a pre-commit
 hook. Furthermore is under very active development and it is quite easy to enable more
checks.

4. Remove deptry.

5. Require poetry 1.2+ for development since we make use of dependency groups.
For example in the sample notebooks we use use `matplotlib` for
visualization and `scipy` for filling missing IOC data. These are **not** `searvey`
dependencies. They are just something we chose to use in our examples. In this sense it
doesn't make sense to have them as "extras", because installing them will not add any
functionality to `searvey`.  Still we want them to be present in the dev environment in order
to be able to develop/use the notebooks. Dependency groups is an elegant way to
both allow installing them and to keep them separated from "core" and "dev" dependencies.

Note: I removed `black` and `ruff` from the dev-dependencies because we can run them
through pre-commit. The benefit of this approach is that these tools run into their own
dedicated environments and thus we have fewer dependencies in the searvey environment.
Running them should be done with `make style` and `make lint` respectively.